requirements.txt
pyttsx3
SpeechRecognition
PyAudio
requests
pygetwindow
rapidfuzz
faster-whisper
gTTS
pygame
pyautogui
Pillow
numpy
silero-vad
piper-tts
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
psutil>=5.9.0
GPUtil>=1.4.0
cryptography>=41.0.0
pyperclip
pywinauto
pywin32
websockets>=12.0
websocket-client>=1.7.0
